[Feat] 대시보드 조치 유형별 건수 및 최근 이상 알림 목록 추가

+++ src/main/java/com/aivle/sellon/domain/channels/service/productmapping/ChannelProductBatchService.java
@@ -74,8 +74,8 @@ public class ChannelProductBatchService {
                         record.get("channel_product_name"),
                         record.isMapped("option_group_names") ? record.get("option_group_names") : null,
                         record.isMapped("channel_option_name") ? record.get("channel_option_name") : null,
-                        parseLongOrNull(record, "sale_price"),
-                        parseLongOrNull(record, "original_price")
+                        parseIntOrNull(record, "sale_price"),
+                        parseIntOrNull(record, "original_price")
                 );
                 count++;
             }
@@ -206,7 +206,7 @@ public class ChannelProductBatchService {
         }
     }
 
-    private Long parseLongOrNull(CSVRecord record, String column) {
+    private Integer parseIntOrNull(CSVRecord record, String column) {
         if (!record.isMapped(column)) {
             return null;
         }
@@ -215,7 +215,7 @@ public class ChannelProductBatchService {
             return null;
         }
         try {
-            return Long.parseLong(value);
+            return Integer.parseInt(value);
         } catch (NumberFormatException e) {
             return null;
         }
